Update AssetLintingSettings class
- Remove trailing spaces
- Move to new namespace
- Update using references
- Reformat code
