+++ Editor/Settings/AssetLintingSettings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 17f4b76b3afb438cb58c855c559d8048
+timeCreated: 1628809239
